refactor: move core command files
