+++ modules/onerepo/src/core/codeowners/__tests__/__fixtures__/repo/modules/burritos/package.json
