+++ modules/onerepo/src/core/tasks/__tests__/__fixtures__/repo/modules/burritos/package.json
