+++ modules/onerepo/src/core/tasks/__tests__/__fixtures__/repo/modules/tacos/package.json
